+++ driver/src/main/scala/org/mongodb/scala/SingleObservable.scala
@@ -1,11 +1,11 @@
 /*
- * Copyright 2015 MongoDB, Inc.
+ * Copyright 2017 MongoDB, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,38 +18,38 @@ package org.mongodb.scala
 
 import com.mongodb.async.client.{Observable => JObservable, Observer => JObserver}
 
-import org.mongodb.scala.internal._
+import org.mongodb.scala.internal.SingleItemObservable
 
 /**
- * A companion object for [[Observable]]
+ * A companion object for [[SingleObservable]]
+ *
+ * @since 2.0
  */
-object Observable {
+object SingleObservable {
 
   /**
-   * Creates an Observable from an Iterable.
+   * Creates an SingleObservable from an item.
    *
    * Convenient for testing and or debugging.
    *
-   * @param from the iterable to create the observable from
-   * @tparam A the type of Iterable
-   * @return an Observable that emits each item from the Iterable
+   * @param item the item to create an observable from
+   * @tparam A the type of the SingleObservable
+   * @return an Observable that emits the item
    */
-  def apply[A](from: Iterable[A]): Observable[A] = IterableObservable[A](from)
+  def apply[A](item: A): SingleObservable[A] = SingleItemObservable(item)
 
 }
 
 /**
- * A `Observable` represents a MongoDB operation.
- *
- * As such it is a provider of a potentially unbounded number of sequenced elements, publishing them according to the demand received
- * from its [[Observer]](s).
+ * A `SingleObservable` represents an [[Observable]] that contains only a single item.
  *
  * @tparam T the type of element signaled.
+ * @since 2.0
  */
-trait Observable[T] extends JObservable[T] {
+trait SingleObservable[T] extends Observable[T] with JObservable[T] {
 
   /**
-   * Request `Observable` to start streaming data.
+   * Request `SingleObservable` to start streaming data.
    *
    * This is a "factory method" and can be called multiple times, each time starting a new [[Subscription]].
    * Each `Subscription` will work for only a single [[Observer]].
@@ -68,3 +68,4 @@ trait Observable[T] extends JObservable[T] {
    */
   override def subscribe(observer: JObserver[_ >: T]): Unit = subscribe(BoxedObserver[T](observer))
 }
+